* I8042prt: Simplify keyboard buffer management.
A ring buffer is absolutely not necessary for the interrupt-DPC
communication and has been simplified down to a linear buffer.

I considered reading the bytes directly when the DPC is run, but
that could be a problem given the buffer size of the i8042 controller.
In the case of a flood of interrupts, that situation probably won't
be very manageable.