* Rtl: Import FreeBSD's tree implementation.
Planning on removing the AVL tree from Boron.  It has had a somewhat
good run, but I can't say it was truly good, because it had some bugs
that haunted me a lot.

I'm sure the original implementation is fine and dandy, but my clone
just wasn't on par.  This should work though :)