* Ke: Fix a bug where, sometimes, timers would simply clash in expiry tick count.
I don't know what causes this as the TSC's resolution is actually very fine grained.
However, because the chance has happened, this simple hack will fix it.

It's a bad hack, but you know what? The chance of it actually happening is like 1/1000
in the fireworks test, so I'll take it!