* Mm: Fix a bug where PTE locations in the recursive map were never TLB-invalidated.
Now, they are, but in the MmGetPteLocation function.  I know this is kind of bad, but
honestly, I don't care too much.  You're not going to get the PTE pointer outside of
MmGetPteLocation / MmGetPteLocationCheck, are you?!

And the cost of invalidating something is minimal anyway. Like, a few clock cycles.

* Mm: Make the page fault handler use PTE pointers from the recursive page map instead of from MiGetPTEPointer.

This makes things slightly faster I think.