* Ke: Add some debugging features to the spin lock.
Optionally, it tracks the RA of the function that acquired a spin lock,
as well as how many spin locks are owned by a particular thread.

Also, attempts to perform a wait while holding a spinlock are checked for.

Also, I forgot to add a change to ki.h in the last commit, which works now.