* Ke: Add support for more aggressive work stealing.
If a processor is overcrowded (so, more than 100 threads are enqueued
to run on it), then other processors will proactively steal from its
work queue.

This currently does not take into account processor affinity (which I
will fix it to do later).

* Ke: Add a feature to track the total time spent in non-idle threads.

This will probably be removed or enhanced at some point, it was useful
when debugging work stealing.