* Ob: Decentralize the object directory mutex.
Now instead of having a central "root directory" mutex, there is a
mutex for each directory.

* Ob: Add ObCreateObjectCallback.

This allows an object to be initialized before it is exposed in the
object namespace.  It prevents race conditions where an object is
exposed prematurely in the object namespace and external code can
access it in its premature state.