* Ke: Add several functions used for "release-and-wait" operations
The list:
	KeReleaseMutexWait
	KeReleaseSemaphoreWait
	KeSetEventWait
	KeResetEventWait
	KePulseEventWait

These functions can be used to create "release-and-wait" atomic building blocks.
For example, POSIX pthreads implement a condition variable object, and one of the
APIs it implements is ``pthread_cond_wait()'' (or its timed variant), which releases
a mutex and waits on a condition variable all in one fell swoop.

Now, I'm not sure these will ACTUALLY get used in userspace, but they're here now
if there'll ever be a need for them.