* Ob: Add a Duplicate method to each object type.
This allows us to create a whole new file object when duplicating
a handle to one, either by duplication or inheritance, rather than
sharing the same file object.

A file object corresponds to an open file description on UNIX like
systems, but unlike the UNIX file systems, this does not track the
individual offset inside the file, rather leaving it up to user
space to do so. However, I think it's still important to keep one
file object per handle.